Move DLL definition to main CKEditor5 package.
