+++ packages/ckeditor5-basic-styles/src/bold.js
@@ -7,7 +7,7 @@
  * @module basic-styles/bold
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import BoldEditing from './bold/boldediting';
 import BoldUI from './bold/boldui';
