+++ packages/ckeditor5-basic-styles/src/bold/boldediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/bold/boldediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const BOLD = 'bold';
