+++ packages/ckeditor5-basic-styles/src/bold/boldui.js
@@ -7,8 +7,8 @@
  * @module basic-styles/bold/boldui
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
-import { ButtonView } from '@ckeditor/ckeditor5-ui/ui';
+import { Plugin } from 'ckeditor5/src/core';
+import { ButtonView } from 'ckeditor5/src/ui';
 
 import boldIcon from '../../theme/icons/bold.svg';
