+++ packages/ckeditor5-basic-styles/src/italic/italicediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/italic/italicediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const ITALIC = 'italic';
