+++ packages/ckeditor5-basic-styles/src/strikethrough/strikethroughediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/strikethrough/strikethroughediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const STRIKETHROUGH = 'strikethrough';
