+++ packages/ckeditor5-basic-styles/src/subscript/subscriptediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/subscript/subscriptediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const SUBSCRIPT = 'subscript';
