+++ packages/ckeditor5-basic-styles/src/subscript/subscriptui.js
@@ -7,8 +7,8 @@
  * @module basic-styles/subscript/subscriptui
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
-import { ButtonView } from '@ckeditor/ckeditor5-ui/ui';
+import { Plugin } from 'ckeditor5/src/core';
+import { ButtonView } from 'ckeditor5/src/ui';
 
 import subscriptIcon from '../../theme/icons/subscript.svg';
