+++ packages/ckeditor5-basic-styles/src/superscript/superscriptediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/superscript/superscriptediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const SUPERSCRIPT = 'superscript';
