+++ packages/ckeditor5-basic-styles/src/underline/underlineediting.js
@@ -7,7 +7,7 @@
  * @module basic-styles/underline/underlineediting
  */
 
-import { Plugin } from '@ckeditor/ckeditor5-core/core';
+import { Plugin } from 'ckeditor5/src/core';
 import AttributeCommand from '../attributecommand';
 
 const UNDERLINE = 'underline';
