+++ packages/ckeditor5-basic-styles/webpack.config.js
@@ -58,16 +58,8 @@ module.exports = {
 	},
 	plugins: [
 		new webpack.DllReferencePlugin( {
-			manifest: require( '../../packages/ckeditor5-dll/build/ckeditor5-dll.manifest.json' ),
-			scope: '@ckeditor/ckeditor5-core'
-		} ),
-		new webpack.DllReferencePlugin( {
-			manifest: require( '../../packages/ckeditor5-dll/build/ckeditor5-dll.manifest.json' ),
-			scope: '@ckeditor/ckeditor5-typing'
-		} ),
-		new webpack.DllReferencePlugin( {
-			manifest: require( '../../packages/ckeditor5-dll/build/ckeditor5-dll.manifest.json' ),
-			scope: '@ckeditor/ckeditor5-ui'
+			manifest: require( '../../build/ckeditor5-dll.manifest.json' ),
+			scope: 'ckeditor5/src'
 		} )
 	]
 };
