+++ sample/index.html
@@ -68,15 +68,15 @@
 <script src="../build/ckeditor5-dll.js"></script>
 
 <!-- Those are stripped-down editors with Essentials plugins loaded from a DLL, exposed on CKEditor global -->
-<script src="../../ckeditor5-dll-classic/build/dll-classic.js"></script>
-<script src="../../ckeditor5-dll-inline/build/dll-inline.js"></script>
-<script src="../../ckeditor5-dll-balloon/build/dll-balloon.js"></script>
-<script src="../../ckeditor5-dll-decoupled-document/build/dll-decoupled-document.js"></script>
+<script src="../packages/ckeditor5-dll-classic/build/dll-classic.js"></script>
+<script src="../packages/ckeditor5-dll-inline/build/dll-inline.js"></script>
+<script src="../packages/ckeditor5-dll-balloon/build/dll-balloon.js"></script>
+<script src="../packages/ckeditor5-dll-decoupled-document/build/dll-decoupled-document.js"></script>
 
 <!-- Editor features as DLL consumer builds exposed as CKEditor5[FeatureName] -->
-<script src="../../ckeditor5-basic-styles/build/basic-styles.js"></script>
-<script src="../../ckeditor5-html-embed/build/html-embed.js"></script>
-<script src="../../ckeditor5-image/build/image.js"></script>
+<script src="../packages/ckeditor5-basic-styles/build/basic-styles.js"></script>
+<script src="../packages/ckeditor5-html-embed/build/html-embed.js"></script>
+<script src="../packages/ckeditor5-image/build/image.js"></script>
 
 <!-- Finally, this is a end-user plugin that uses DLL -->
 <script src="../build/dll-consumer-plugin.js"></script>
