+++ sample/plugin/src/dllconsumerplugin.js
