+++ src/classiceditor.js
