+++ src/decouplededitor.js
