+++ src/inlineeditor.js
