+++ src/selectall.js
