+++ src/typing.js
